+++ Assets/Scripts/Data/BuildingData/BuildingData.cs
@@ -10,6 +10,6 @@ public abstract class BuildingData : ScriptableObject
 
 	public abstract float GetCycleSeconds(int level = 0);
 	public abstract int GetCatchAmount(int level);
-	public abstract int GetUpgradeCost(int level, int indexMultiplier);
-	public abstract int GetBuildCost(int indexMultiplier);
+	public abstract long GetUpgradeCost(int level, int indexMultiplier);
+	public abstract long GetBuildCost(int indexMultiplier);
 }
